pyproject.toml
[project]
name = "ensemble-cache-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "redis>=5.0.1",
    "fastapi>=0.115.12",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.2",
]
